change file extensions
